Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0662-maximum-width-of-binary-tree/0662-maximum-width-of-binary-tree.cpp
@@ -15,7 +15,7 @@ public:
         if(root==NULL)return 0;
         queue<pair<TreeNode*, int>>q;
         q.push({root, 0});
-        int maxw = 0;
+        long long maxw = 0;
         while(!q.empty())
         {
             auto topo = q.front();
